Move the OptionTypeMap code closer to home
This is a predecessor to the Configurable PR.  This change moves the OptionTypeInfo maps closer to where they will be used.

Additionally, the "is_mutable" boolean is changed to an OptionTypeFlags to allow for future expansion/additional use of this field.

When the Configurable changes are adopted, these values will become static and not associated with the OptionsHelper.