Add ObjectRegistry to DBOptions
The ObjectRegistry now has two new subclasses of libraries. One class is "local", which is primarily used for testing. Local libraries allow another named library to be put into an object registry context.

The other type is "dynamic", which allows a DynamicLibrary to be registered with the ObjectRegistry. The registration also registers a function in that library to invoke to populate the object library registry. The dynamic registry library can also be serialized to/from the INI file, meaning that once one is registered by name/stirng in the INI file