Add a CloudStorageProvider class to CloudEnv
This change moves the functionality for dealing with S3-like bucket objects  into a StorageProvider class.  The CloudEnvOptions now has a CloudStorageProvider instance.

Additionally, the code was "cleaned up" to allow it to more easily be written as "Configurable" and "Customizable".  This basically means that the construction of objects was separated from the configuration of objects was separated from the verification/preparation of objects.  By separating these three steps, the objects (AwsEnv, S3StorageProvider, Kinesis and Kafka LogControllers) will be easily adapted to loading from property/INI files.