Add an option to use constant file size in SSTFileManagerImpl
Summary:
The most expensive operations during DB opening in GHOST mode is checking the size of all SST files for bookkeeping in SSTFileManager. This patch adds an option to use a constant size for SSTFileManager instead.
Changes in `SSTFileManagerImpl` are already made upstream in https://github.com/facebook/rocksdb/pull/6363, so it shouldn't be too bad when we merge.

Test Plan: Add a unit test.

Reviewers: dhruba, igor

Differential Revision: https://rockset.phacility.com/D5613