Better support for shared libraries in RocksDB
Adds the ability to build RocksDB tests and tools in shared library mode.  Change the linking of tests/tools to be against a library rather than a list of objects.  This change substantially reduces the size of the objects produced.

Added LIB_MODE option to Makefile to allow "static" or "shared" builds.

Added OBJ_DIR option to Makefile to allow objects to be placed in an alternative location.  By default, OBJ_DIR is objects/$LIB_MODE.