Options to disable GetChildren from cloud
Summary: Add an option that short-circuits the `GetChildren` call in the cloud. This option is useful during DB Opening where `GetChildren` becomes too expensive, since it was called multiple times. Only use this when opening DB in a temporary clone.

Test Plan: Run in my namespace. This cuts the DB Opening time from 7s to < 800ms

Reviewers: dhruba

Differential Revision: https://rockset.phacility.com/D6035